yay it works! Only on my machine

+++ src/fr/entasia/factools/Utils.java
@@ -2,8 +2,8 @@ package fr.entasia.factools;
 
 public class Utils {
 
-	public static int getRandom(int min, int max) {
-		return (int) ((Math.random() * (max-min)) + min);
+	public static double getRandom(double min, double max) {
+		return ((Math.random() * (max-min)) + min);
 	}
 
 }
